Representations now use molecule class.
Molecule class need to be slightly modified to allow use to "overload" the constructor.
We can call a null constructor and just make the molecule the instantiate.
Or we can call the constructor with a filename and it will create and instatiate the molecule.

Both representations now use the molecule class. Previous the code checked if the file was
of a supported type before making the rep. This has been removed. I think that this should be
handled by the molecule class followed by a try,except,pass. I will open an issue for discussion
and we should probably discuss in person as well.